backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
pytest==7.4.3
hypothesis==6.92.1
cyborgdb==0.14.0
cyborgdb_core==0.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sentence-transformers==2.7.0
PyPDF2==3.0.1
python-docx==1.1.0
httpx==0.25.2
pytest-asyncio==0.21.1
python-dotenv==1.0.0